Move GetHoverAsync tests to Workspaces
Moving the existing hover tests requires a few test infrastructure changes.
